+++ deps.sh
@@ -27,7 +27,7 @@ rm "${LUACOV_COVERALLS_ROCKSPEC_FILE}"
 rmdir "${TMPDIR}"
 
 tarantoolctl rocks install cartridge 2.7.4
-tarantoolctl rocks install ddl 1.6.0
+tarantoolctl rocks install ddl 1.6.2
 tarantoolctl rocks install migrations 0.4.2
 
 tarantoolctl rocks make